requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.11
flask-cors==6.0.1
python-dotenv==1.2.1







